extract checkAdmin unto reusable hook
